fixing issue related to missing attributes
