+++ quspin/basis/basis_1d/base_1d.py
@@ -44,7 +44,7 @@ class bitops:
 
 class basis_1d(lattice_basis):
 	def __init__(self,basis_module,ops_module,L,Np=None,pars=None,count_particles=False,**blocks):
-
+		lattice_basis.__init__(self)
 		if self.__class__.__name__ == "basis_1d":
 			raise ValueError("This class is not intended"
 							 " to be instantiated directly.")
