+++ quspin/basis/basis_general/base_general.py
@@ -4,17 +4,7 @@ import os,numexpr
 from ._basis_general_core.general_basis_utils import basis_int_to_python_int,_get_basis_index
 from ._basis_general_core import basis_zeros
 from ..lattice import lattice_basis
-from numba import njit
 import warnings
-
-
-@njit
-def _is_sorted_decending(a):
-	for i in range(a.size-1):
-		if(a[i]<a[i+1]):
-			return False
-
-	return True
 
 class GeneralBasisWarning(Warning):
 	pass
@@ -87,6 +77,7 @@ def check_symmetry_maps(item1,item2):
 
 class basis_general(lattice_basis):
 	def __init__(self,N,block_order=None,**kwargs):
+		lattice_basis.__init__(self)
 		self._unique_me = True
 		self._check_herm = True
 
@@ -766,23 +757,20 @@ class basis_general(lattice_basis):
 		if Ns < 0:
 				raise ValueError("estimate for size of reduced Hilbert-space is too low, please double check that transformation mappings are correct or use 'Ns_block_est' argument to give an upper bound of the block size.")
 
-		# # sort basis
-		# if type(self._Np) is int or type(self._Np) is tuple or self._Np is None:
-		# 	if Ns > 0:
-		# 		# self._basis = basis[Ns-1::-1].copy()
-		# 		# self._n = n[Ns-1::-1].copy()
-		# 		# if Np_list is not None: self._Np_list = Np_list[Ns-1::-1].copy()
-		# 		self._basis = basis[:Ns].copy()
-		# 		self._n = n[:Ns].copy()
-		# 		if Np_list is not None: self._Np_list = Np_list[:Ns].copy()
-		# 	else:
-		# 		self._basis = _np.array([],dtype=basis.dtype)
-		# 		self._n = _np.array([],dtype=n.dtype)
-		# 		if Np_list is not None: self._Np_list = _np.array([],dtype=Np_list.dtype)
+		# sort basis
+		if type(self._Np) is int or type(self._Np) is tuple or self._Np is None:
+			if Ns > 0:
+				self._basis = basis[:Ns].copy()
+				self._n = n[:Ns].copy()
+				if Np_list is not None: self._Np_list = Np_list[:Ns].copy()
+			else:
+				self._basis = _np.array([],dtype=basis.dtype)
+				self._n = _np.array([],dtype=n.dtype)
+				if Np_list is not None: self._Np_list = _np.array([],dtype=Np_list.dtype)
 
-		# 	sort_basis = False
-		# else:
-		# 	sort_basis = True
+			sort_basis = False
+		else:
+			sort_basis = True
 
 		if Ns > 0:
 			# self._basis = basis[Ns-1::-1].copy()
@@ -791,12 +779,6 @@ class basis_general(lattice_basis):
 			self._basis = basis[:Ns].copy()
 			self._n = n[:Ns].copy()
 			if Np_list is not None: self._Np_list = Np_list[:Ns].copy()
-
-			if not _is_sorted_decending(self._basis):
-				ind = _np.argsort(self._basis,kind="heapsort")[::-1]
-				self._basis = _np.asarray(self._basis[ind],order="C")
-				self._n =  _np.asarray(self._n[ind],order="C")
-				if Np_list is not None: self._Np_list = _np.asarray(self._Np_list[ind],order="C")
 		else:
 			self._basis = _np.array([],dtype=basis.dtype)
 			self._n = _np.array([],dtype=n.dtype)
